requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.1.6
django-bootstrap5==25.1
django-cleanup==9.0.0
django-js-asset==3.1.0
django-tinymce==4.1.0
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
webencodings==0.5.1
